organize Preview under preview directory
